Split ImportServiceHandle into per-area files
handle.rs had grown to ~1,961 lines mixing a wide command/query API with
its helpers and ~870 lines of inline tests. Relocate the impl methods into
per-area files (search, scan, import, watch) under import/handle/, keeping
the types, channel wiring, constructor, and shared helpers in mod.rs. Pure
relocation: no signature or behavior changes. `super::` module paths in the
moved methods are requalified to absolute `crate::import::` paths to account
for the deeper nesting.
